Extract useCancellation to another file
